# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Blog publico",
        "description": "API para blog de la prueba técnica para koltin",
        "contact": {
            "email": "daniel.rpda@gmail.com"
        },
        "version": "1.0.0",
        "x-logo": {
            "url": "https://via.placeholder.com/190x90.png?text=L5-Swagger"
        }
    },
    "paths": {
        "/api/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login endpoint to get token and allow to create posts and comments",
                "operationId": "ea134134deae11eb974c719450ef8a02",
                "requestBody": {
                    "description": "Provide All Info Below",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "email@test.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "12345678"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User login successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "User Logged in successfully"
                                        },
                                        "token": {
                                            "type": "string",
                                            "example": "QrCcHEPyOy66h7ibLWaZZ7I"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "type": "number",
                                            "example": 422
                                        },
                                        "message": {
                                            "type": "object",
                                            "example": "{email:[required]}"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Required authentication",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "type": "number",
                                            "example": 401
                                        },
                                        "message": {
                                            "type": "object",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/posts": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "Paginated post list with user and post",
                "operationId": "fbf49a8a37f41e6323a594a41dd0d7cd",
                "responses": {
                    "200": {
                        "description": "Author registered successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "current_page": {
                                            "type": "number",
                                            "example": 1
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "number",
                                                        "example": 1
                                                    },
                                                    "title": {
                                                        "type": "string",
                                                        "example": "example"
                                                    },
                                                    "text": {
                                                        "type": "string",
                                                        "example": "example"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "first_page_url": {
                                            "type": "string",
                                            "example": "http://127.0.0.1/api/posts?page=1"
                                        },
                                        "from": {
                                            "type": "number",
                                            "example": 1
                                        },
                                        "last_page": {
                                            "type": "number",
                                            "example": 1
                                        },
                                        "last_page_url": {
                                            "type": "string",
                                            "example": "http://127.0.0.1/api/posts?page=1"
                                        },
                                        "per_page": {
                                            "type": "number",
                                            "example": 1
                                        },
                                        "prev_page_url": {
                                            "type": "string",
                                            "example": "http://127.0.0.1/api/posts?page=1"
                                        },
                                        "to": {
                                            "type": "number",
                                            "example": 1
                                        },
                                        "total": {
                                            "type": "number",
                                            "example": 1
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Required authentication",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "type": "number",
                                            "example": 401
                                        },
                                        "message": {
                                            "type": "object",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Posts"
                ],
                "summary": "Register a new post",
                "operationId": "43be35d46510ea3fcbe12fa9a92d1eff",
                "requestBody": {
                    "description": "Provide All Info Below",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "name",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "Test Post"
                                    },
                                    "text": {
                                        "type": "text",
                                        "format": "text",
                                        "example": "new post"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Post registered successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Post created successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Post Not register",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "type": "number",
                                            "example": 422
                                        },
                                        "message": {
                                            "type": "object",
                                            "example": "{email:[required]}"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Required authentication",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "type": "number",
                                            "example": 401
                                        },
                                        "message": {
                                            "type": "object",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/authors": {
            "get": {
                "tags": [
                    "Authors"
                ],
                "summary": "Paginated authors list with comments and post",
                "operationId": "aef2397961a0e3fdfc433aa36b8707b2",
                "responses": {
                    "200": {
                        "description": "Author registered successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "current_page": {
                                            "type": "number",
                                            "example": 1
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "number",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "example"
                                                    },
                                                    "email": {
                                                        "type": "string",
                                                        "example": "example"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "first_page_url": {
                                            "type": "string",
                                            "example": "http://127.0.0.1/api/posts?page=1"
                                        },
                                        "from": {
                                            "type": "number",
                                            "example": 1
                                        },
                                        "last_page": {
                                            "type": "number",
                                            "example": 1
                                        },
                                        "last_page_url": {
                                            "type": "string",
                                            "example": "http://127.0.0.1/api/posts?page=1"
                                        },
                                        "per_page": {
                                            "type": "number",
                                            "example": 1
                                        },
                                        "prev_page_url": {
                                            "type": "string",
                                            "example": "http://127.0.0.1/api/posts?page=1"
                                        },
                                        "to": {
                                            "type": "number",
                                            "example": 1
                                        },
                                        "total": {
                                            "type": "number",
                                            "example": 1
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Authors"
                ],
                "summary": "Register a new author",
                "operationId": "047111b702e7fabdca63136418df66c3",
                "requestBody": {
                    "description": "Provide All Info Below",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "name",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "Mario"
                                    },
                                    "email": {
                                        "type": "email",
                                        "format": "text",
                                        "example": "mario@example.org"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "123467890"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "123467890"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Author registered successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "false"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Author registered successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Author Not register",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "type": "number",
                                            "example": 422
                                        },
                                        "message": {
                                            "type": "object",
                                            "example": "message:{email:[required]}"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}